Further progress on model structure
Added and expanded the functions from the old model. added a variables file, which should encode program variables, currently only the default device of the pytorch tensors. Next update should move some classes out of the notebook, and into a seperate file, and should see the start of the implementation of the rate-equation model.