Update expanding classes and moving them to seperate file.
This is an update to the particles and laserbeam class, allowing for the calculation of the intensity of a given laserbeam at some set of points. And heavily upgrading the particles class to allow for the addition of new particles based on pressure and timestep size. Some minor demonstrators have been left in the main notebook, to show how these functions can be used.